moved SteppyError to separate py file
